refactor: 결제 도메인 로직 확장으로 인해 패키지 구조 정리
