+++ apps/commerce-api/src/main/java/com/loopers/domain/payment/PaymentAdapter.java
@@ -1,6 +1,6 @@
 package com.loopers.domain.payment;
 
-import com.loopers.infrastructure.payment.PgClient;
+import com.loopers.infrastructure.payment.client.PgClient;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
