--- apps/commerce-api/src/main/java/com/loopers/domain/payment/PaymentProcessor.java
@@ -1,6 +1,0 @@
-package com.loopers.domain.payment;
-
-public interface PaymentProcessor {
-
-    Payment pay(PaymentCommand paymentCommand);
-}
