+++ apps/commerce-api/src/main/java/com/loopers/domain/payment/processor/CardPaymentProcessor.java
@@ -1,6 +1,7 @@
-package com.loopers.domain.payment;
+package com.loopers.domain.payment.processor;
 
-import com.loopers.infrastructure.payment.PgClient;
+import com.loopers.domain.payment.*;
+import com.loopers.infrastructure.payment.client.PgClient;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
