+++ apps/commerce-api/src/main/java/com/loopers/domain/payment/processor/PaymentProcessor.java
@@ -1,0 +1,9 @@
+package com.loopers.domain.payment.processor;
+
+import com.loopers.domain.payment.Payment;
+import com.loopers.domain.payment.PaymentCommand;
+
+public interface PaymentProcessor {
+
+    Payment pay(PaymentCommand paymentCommand);
+}
