+++ apps/commerce-api/src/main/java/com/loopers/infrastructure/payment/repository/PaymentCoreRepository.java
@@ -1,4 +1,4 @@
-package com.loopers.infrastructure.payment;
+package com.loopers.infrastructure.payment.repository;
 
 import com.loopers.domain.payment.Payment;
 import com.loopers.domain.payment.PaymentRepository;
