+++ apps/commerce-api/src/main/java/com/loopers/infrastructure/payment/repository/PaymentEntity.java
@@ -1,4 +1,4 @@
-package com.loopers.infrastructure.payment;
+package com.loopers.infrastructure.payment.repository;
 
 import com.loopers.domain.BaseEntity;
 import com.loopers.domain.payment.Payment;
